DB_AI/requirements.txt
vanna>=0.7.0
llama-cpp-python>=0.2.90
chromadb>=0.4.22
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
huggingface-hub>=0.20.0
requests>=2.31.0
tqdm>=4.66.0
